requirements.txt
transformers==4.35.2
torch==2.1.0
pinecone-client[grpc]==2.2.1


